Moving SQSOptions
