+++ src/SQSConsumerOptions.ts
@@ -1,5 +1,5 @@
 import { LoggerOptions } from './logger/LoggerOptions';
-import { SQSOptions } from './SQSOptions';
+import { SQSOptions } from './sqs/SQSOptions';
 
 export type SQSConsumerOptions = {
   logOptions?: LoggerOptions;
