+++ src/sqs/SQSOptions.ts
